scr/firstproject/requirements.txt
asgiref==3.3.4
Django==3.2.3
django-cors-headers==3.7.0
djangorestframework==3.12.4
django-crontab==0.7.1
gunicorn==20.1.0
lxml==4.6.3
numpy==1.21.0
opencv-python==4.5.2.54
Pillow==8.3.0
pytesseract==0.3.8
python-docx==0.8.11
pytz==2021.1
requests==2.26.0
sqlparse==0.4.1
num2words==0.5.10
